Move Swift/C crypto wrappers to tests
Unused in the library, except for ZeroingData.

Fix outdated bridged ObjC implementations
