+++ Tests/OpenVPN/CryptoOpenSSL/Native/CryptoAEAD.swift
@@ -1,5 +1,5 @@
 //
-//  CCryptoCBC.swift
+//  CryptoAEAD.swift
 //  Partout
 //
 //  Created by Davide De Rosa on 6/14/25.
@@ -26,44 +26,45 @@
 internal import _PartoutCryptoOpenSSL_C
 import Foundation
 
-public final class CryptoCBC: Encrypter, Decrypter {
-    private let ptr: UnsafeMutablePointer<crypto_cbc_t>
+final class CryptoAEAD: Encrypter, Decrypter {
+    private let ptr: UnsafeMutablePointer<crypto_aead_t>
 
     private let mappedError: (CryptoError) -> Error
 
-    public init(
-        cipherName: String?,
-        digestName: String,
+    init(
+        cipherName: String,
+        tagLength: Int,
+        idLength: Int,
         mappedError: ((CryptoError) -> Error)? = nil
     ) throws {
-        guard let ptr = crypto_cbc_create(cipherName, digestName) else {
+        guard let ptr = crypto_aead_create(cipherName, tagLength, idLength) else {
             throw CryptoError()
         }
-        NSLog("PartoutOpenVPN: Using CryptoCBC (Swift)")
+        NSLog("PartoutOpenVPN: Using CryptoAEAD (Swift)")
         self.ptr = ptr
         self.mappedError = mappedError ?? { $0 }
     }
 
-    public var digestLength: Int {
+    var digestLength: Int {
         ptr.pointee.crypto.meta.digest_len
     }
 
-    public var tagLength: Int {
+    var tagLength: Int {
         ptr.pointee.crypto.meta.tag_len
     }
 
-    public func encryptionCapacity(for length: Int) -> Int {
+    func encryptionCapacity(for length: Int) -> Int {
         ptr.pointee.crypto.meta.encryption_capacity(ptr, length)
     }
 
-    public func configureEncryption(withCipherKey cipherKey: ZeroingData?, hmacKey: ZeroingData?) {
-        guard let hmacKey else {
-            fatalError("HMAC key required")
+    func configureEncryption(withCipherKey cipherKey: ZeroingData?, hmacKey: ZeroingData?) {
+        guard let cipherKey, let hmacKey else {
+            return
         }
-        ptr.pointee.crypto.encrypter.configure(ptr, cipherKey?.ptr, hmacKey.ptr)
+        ptr.pointee.crypto.encrypter.configure(ptr, cipherKey.ptr, hmacKey.ptr)
     }
 
-    public func encryptBytes(_ bytes: UnsafePointer<UInt8>, length: Int, dest: UnsafeMutablePointer<UInt8>, destLength: UnsafeMutablePointer<Int>, flags: CryptoFlagsWrapper?) throws -> Bool {
+    func encryptBytes(_ bytes: UnsafePointer<UInt8>, length: Int, dest: UnsafeMutablePointer<UInt8>, destLength: UnsafeMutablePointer<Int>, flags: CryptoFlagsWrapper?) throws -> Bool {
         var code = CryptoErrorGeneric
         var cFlags = crypto_flags_t()
         let flagsPtr = flags.pointer(to: &cFlags)
@@ -73,14 +74,14 @@ public final class CryptoCBC: Encrypter, Decrypter {
         return true
     }
 
-    public func configureDecryption(withCipherKey cipherKey: ZeroingData?, hmacKey: ZeroingData?) {
-        guard let hmacKey else {
-            fatalError("HMAC key required")
+    func configureDecryption(withCipherKey cipherKey: ZeroingData?, hmacKey: ZeroingData?) {
+        guard let cipherKey, let hmacKey else {
+            return
         }
-        ptr.pointee.crypto.decrypter.configure(ptr, cipherKey?.ptr, hmacKey.ptr)
+        ptr.pointee.crypto.decrypter.configure(ptr, cipherKey.ptr, hmacKey.ptr)
     }
 
-    public func decryptBytes(_ bytes: UnsafePointer<UInt8>, length: Int, dest: UnsafeMutablePointer<UInt8>, destLength: UnsafeMutablePointer<Int>, flags: CryptoFlagsWrapper?) throws -> Bool {
+    func decryptBytes(_ bytes: UnsafePointer<UInt8>, length: Int, dest: UnsafeMutablePointer<UInt8>, destLength: UnsafeMutablePointer<Int>, flags: CryptoFlagsWrapper?) throws -> Bool {
         var code = CryptoErrorGeneric
         var cFlags = crypto_flags_t()
         let flagsPtr = flags.pointer(to: &cFlags)
@@ -90,11 +91,7 @@ public final class CryptoCBC: Encrypter, Decrypter {
         return true
     }
 
-    public func verifyBytes(_ bytes: UnsafePointer<UInt8>, length: Int, flags: CryptoFlagsWrapper? = nil) throws -> Bool {
-        var code = CryptoErrorGeneric
-        guard ptr.pointee.crypto.decrypter.verify(ptr, bytes, length, &code) else {
-            throw mappedError(CryptoError(code))
-        }
-        return true
+    func verifyBytes(_ bytes: UnsafePointer<UInt8>, length: Int, flags: CryptoFlagsWrapper?) throws -> Bool {
+        fatalError("Unsupported")
     }
 }
