+++ Tests/OpenVPN/CryptoOpenSSL/Native/CryptoCTR.swift
@@ -1,5 +1,5 @@
 //
-//  CryptoAEAD.swift
+//  CCryptoCTR.swift
 //  Partout
 //
 //  Created by Davide De Rosa on 6/14/25.
@@ -26,45 +26,46 @@
 internal import _PartoutCryptoOpenSSL_C
 import Foundation
 
-public final class CryptoAEAD: Encrypter, Decrypter {
-    private let ptr: UnsafeMutablePointer<crypto_aead_t>
+final class CryptoCTR: Encrypter, Decrypter {
+    private let ptr: UnsafeMutablePointer<crypto_ctr_t>
 
     private let mappedError: (CryptoError) -> Error
 
-    public init(
+    init(
         cipherName: String,
+        digestName: String,
         tagLength: Int,
-        idLength: Int,
+        payloadLength: Int,
         mappedError: ((CryptoError) -> Error)? = nil
     ) throws {
-        guard let ptr = crypto_aead_create(cipherName, tagLength, idLength) else {
+        guard let ptr = crypto_ctr_create(cipherName, digestName, tagLength, payloadLength) else {
             throw CryptoError()
         }
-        NSLog("PartoutOpenVPN: Using CryptoAEAD (Swift)")
+        NSLog("PartoutOpenVPN: Using CryptoCTR (Swift)")
         self.ptr = ptr
         self.mappedError = mappedError ?? { $0 }
     }
 
-    public var digestLength: Int {
+    var digestLength: Int {
         ptr.pointee.crypto.meta.digest_len
     }
 
-    public var tagLength: Int {
+    var tagLength: Int {
         ptr.pointee.crypto.meta.tag_len
     }
 
-    public func encryptionCapacity(for length: Int) -> Int {
+    func encryptionCapacity(for length: Int) -> Int {
         ptr.pointee.crypto.meta.encryption_capacity(ptr, length)
     }
 
-    public func configureEncryption(withCipherKey cipherKey: ZeroingData?, hmacKey: ZeroingData?) {
+    func configureEncryption(withCipherKey cipherKey: ZeroingData?, hmacKey: ZeroingData?) {
         guard let cipherKey, let hmacKey else {
             return
         }
         ptr.pointee.crypto.encrypter.configure(ptr, cipherKey.ptr, hmacKey.ptr)
     }
 
-    public func encryptBytes(_ bytes: UnsafePointer<UInt8>, length: Int, dest: UnsafeMutablePointer<UInt8>, destLength: UnsafeMutablePointer<Int>, flags: CryptoFlagsWrapper?) throws -> Bool {
+    func encryptBytes(_ bytes: UnsafePointer<UInt8>, length: Int, dest: UnsafeMutablePointer<UInt8>, destLength: UnsafeMutablePointer<Int>, flags: CryptoFlagsWrapper?) throws -> Bool {
         var code = CryptoErrorGeneric
         var cFlags = crypto_flags_t()
         let flagsPtr = flags.pointer(to: &cFlags)
@@ -74,14 +75,14 @@ public final class CryptoAEAD: Encrypter, Decrypter {
         return true
     }
 
-    public func configureDecryption(withCipherKey cipherKey: ZeroingData?, hmacKey: ZeroingData?) {
+    func configureDecryption(withCipherKey cipherKey: ZeroingData?, hmacKey: ZeroingData?) {
         guard let cipherKey, let hmacKey else {
             return
         }
         ptr.pointee.crypto.decrypter.configure(ptr, cipherKey.ptr, hmacKey.ptr)
     }
 
-    public func decryptBytes(_ bytes: UnsafePointer<UInt8>, length: Int, dest: UnsafeMutablePointer<UInt8>, destLength: UnsafeMutablePointer<Int>, flags: CryptoFlagsWrapper?) throws -> Bool {
+    func decryptBytes(_ bytes: UnsafePointer<UInt8>, length: Int, dest: UnsafeMutablePointer<UInt8>, destLength: UnsafeMutablePointer<Int>, flags: CryptoFlagsWrapper?) throws -> Bool {
         var code = CryptoErrorGeneric
         var cFlags = crypto_flags_t()
         let flagsPtr = flags.pointer(to: &cFlags)
@@ -91,7 +92,7 @@ public final class CryptoAEAD: Encrypter, Decrypter {
         return true
     }
 
-    public func verifyBytes(_ bytes: UnsafePointer<UInt8>, length: Int, flags: CryptoFlagsWrapper?) throws -> Bool {
+    func verifyBytes(_ bytes: UnsafePointer<UInt8>, length: Int, flags: CryptoFlagsWrapper?) throws -> Bool {
         fatalError("Unsupported")
     }
 }
