+++ Tests/OpenVPN/CryptoOpenSSL/ZeroingDataTests.swift
@@ -23,7 +23,7 @@
 //  along with Partout.  If not, see <http://www.gnu.org/licenses/>.
 //
 
-import _PartoutCryptoOpenSSL
+internal import _PartoutCryptoOpenSSL
 #if canImport(_PartoutCryptoOpenSSL_ObjC)
 internal import _PartoutCryptoOpenSSL_ObjC
 #endif
